src/requirements.txt
absl-py==1.4.0
alabaster==0.7.13
argh==0.28.1
astunparse==1.6.3
Babel==2.12.1
blis==0.7.9
bumpversion==0.5.3
cachetools==5.3.0
catalogue==1.0.2
certifi==2020.12.5
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clickhouse-connect==0.5.25
clickhouse-driver==0.2.6
confection==0.0.4
contourpy==1.0.7
coverage==4.1
cryptography==3.3.2
cycler==0.11.0
cymem==2.0.7
distlib==0.3.6
dlib==19.24.1
docopt==0.6.2
docutils==0.19
emoji==0.5.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz#sha256=06c80936324012d1223291d2af41a5229e746dc2dee8fe31a532666ee3d18aaa
face-recognition==1.4.0
face-recognition-models==0.3.0
filelock==3.12.0
flake8==6.0.0
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
imageio==2.28.1
imagesize==1.4.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
imutils==0.5.4
jax==0.4.8
Jinja2==3.0.3
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
libclang==16.0.0
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.7.0
ml-dtypes==0.1.0
mpmath==1.3.0
msgpack==0.5.6
murmurhash==1.0.9
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.5.1.48
openface==0.0.0
opt-einsum==3.3.0
packaging==23.1
paho-mqtt==1.6.1
pandas==2.0.1
pathtools==0.1.2
pathy==0.10.1
Pillow==9.5.0
pipreqs==0.4.13
plac==1.1.3
platformdirs==3.5.0
pluggy==0.3.1
preshed==3.0.8
protobuf==4.22.4
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycld2==0.41
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyflakes==3.0.1
Pygments==2.15.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
QDarkStyle==3.1
QtPy==2.3.1
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
six==1.16.0
smart-open==6.3.0
snowballstemmer==2.2.0
spacy==2.3.9
spacy-cld==0.1.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
Sphinx==1.6.7
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
srsly==1.0.6
sympy==1.11.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
thinc==7.4.6
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.1
torchtext==0.3.1
tox==2.3.1
tqdm==4.65.0
tweet-preprocessor==0.6.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.15
virtualenv==20.23.0
wasabi==0.10.1
watchdog==0.8.3
Werkzeug==2.3.3
wincertstore==0.2
wrapt==1.14.1
yarg==0.1.9
zipp==3.15.0
zstandard==0.21.0